requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.31.1
SQLAlchemy==2.0.36
pyodbc==5.2.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
